Refactor Custom Link Storybook component to share in multiple storybooks
